+++ lib/core/network/auth_interceptor.dart
@@ -1,0 +1,13 @@
+import 'package:dio/dio.dart';
+import 'package:payout/features/auth/services/session_manager.dart';
+
+class AuthInterceptor extends Interceptor {
+  @override
+  void onRequest(RequestOptions options, RequestInterceptorHandler handler) {
+    final token = SessionManager.instance.accessToken;
+    if (token != null) {
+      options.headers['Authorization'] = 'Bearer $token';
+    }
+    super.onRequest(options, handler);
+  }
+}