+++ lib/core/network/logging_interceptor.dart
@@ -1,0 +1,22 @@
+import 'package:dio/dio.dart';
+import 'package:flutter/foundation.dart';
+
+class LoggingInterceptor extends Interceptor {
+  @override
+  void onRequest(RequestOptions options, RequestInterceptorHandler handler) {
+    if (kDebugMode) {
+      print('--> ${options.method} ${options.uri}');
+      print('Headers: ${options.headers}');
+    }
+    super.onRequest(options, handler);
+  }
+
+  @override
+  void onResponse(Response response, ResponseInterceptorHandler handler) {
+    if (kDebugMode) {
+      print('<-- ${response.statusCode} ${response.requestOptions.uri}');
+      print('Response: ${response.data}');
+    }
+    super.onResponse(response, handler);
+  }
+}