+++ lib/core/security/biometric_manager.dart
@@ -1,0 +1,9 @@
+class BiometricManager {
+  static Future<bool> isBiometricsAvailable() async {
+    return true;
+  }
+
+  static Future<bool> authenticate() async {
+    return true;
+  }
+}