+++ test/mock/mock_test_utils.dart
@@ -1,0 +1,9 @@
+class MockTestUtils {
+  static Map<String, dynamic> getMockUserData() {
+    return {
+      'id': 'usr_9091',
+      'name': 'John Doe',
+      'email': 'john.doe@example.com',
+    };
+  }
+}